+++ Sources/Sentry/Public/SentryOptions.h
@@ -114,6 +114,11 @@ NS_SWIFT_NAME(Options)
 @property (nonatomic, assign) BOOL enableAutoSessionTracking;
 
 /**
+ * Whether to enable to enable out of memory tracking or not. Default is YES.
+ */
+@property (nonatomic, assign) BOOL enableOutOfMemoryTracking;
+
+/**
  * The interval to end a session if the App goes to the background.
  */
 @property (nonatomic, assign) NSUInteger sessionTrackingIntervalMillis;