+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- feat: Async callstacks are being tracked by wrapping the `dispatch_async` and related APIs. #998
+
 ## 7.0.0-alpha.3
 
 - feat: Out Of Memory Tracking #1001